Allow login using email address - first steps
Make changes to login screen and function so that the user's primary
email address is acceptable in place of username. Appears to work OK
both in main login page and in plugin, but we need to change
the rules about usernames so they can't possibly collide with an
email address.